requirements.txt
streamlit
gspread
oauth2client
pandas
google-cloud-secret-manager
google-auth                  